[asterixdb] Updated the upload.sh script to reduce the amount of data uploaded.

+++ experiments/asterixdb-sqlpp/upload.sh
@@ -1,6 +1,7 @@
 #!/usr/bin/env bash
 
 SCRIPT_PATH="$( cd "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
+ENABLE_JSON_DATA_COPY=false
 
 # Load common functions
 . "$SCRIPT_PATH/../common/ec2-helpers.sh"
@@ -27,23 +28,25 @@ echo "Uploading data..."
 
     for n in $(for l in {0..16}; do echo $((2**$l*1000)); done) $NSF1
     do
-        dataset_name="Run2012B_SingleMu_restructured_json_${n}"
-
         # Copy JSON data from S3 to HDFS
-        ssh -q ec2-user@${dnsnames[0]} \
-            <<-EOF
-			mkdir -p /data/tmp
-			aws s3 cp --no-progress --recursive "$S3_INPUT_PATH/$dataset_name/" "/data/tmp/$dataset_name/"
-			./hadoop/bin/hadoop fs -mkdir "/$dataset_name"
-			./hadoop/bin/hadoop fs -put "/data/tmp/$dataset_name/" /
-			rm -r "/data/tmp/$dataset_name/"
-			EOF
+        if [ "${ENABLE_JSON_DATA_COPY}" = true ]; then
+            dataset_name="Run2012B_SingleMu_restructured_json_${n}"
 
-        for dnsname in ${dnsnames[@]}
-        do
-            ssh -q ec2-user@$dnsname \
-                aws s3 cp --no-progress --recursive "$S3_INPUT_PATH/$dataset_name/" "/data/input/$dataset_name/"
-        done
+            ssh -q ec2-user@${dnsnames[0]} \
+                <<-EOF
+				mkdir -p /data/tmp
+				aws s3 cp --no-progress --recursive "$S3_INPUT_PATH/$dataset_name/" "/data/tmp/$dataset_name/"
+				./hadoop/bin/hadoop fs -mkdir "/$dataset_name"
+				./hadoop/bin/hadoop fs -put "/data/tmp/$dataset_name/" /
+				rm -r "/data/tmp/$dataset_name/"
+				EOF
+
+            for dnsname in ${dnsnames[@]}
+            do
+                ssh -q ec2-user@$dnsname \
+                    aws s3 cp --no-progress --recursive "$S3_INPUT_PATH/$dataset_name/" "/data/input/$dataset_name/"
+            done
+        fi
 
         dataset_name="Run2012B_SingleMu_restructured_${n}"
 
@@ -86,79 +89,81 @@ echo "Uploading data..."
         # JSON
         #
 
-        dataset_name="Run2012B_SingleMu_restructured_json_${n}"
+        if [ "${ENABLE_JSON_DATA_COPY}" = true ]; then 
+            dataset_name="Run2012B_SingleMu_restructured_json_${n}"
 
-        # Internal relations loaded from files on HDFS
-        "$SCRIPT_PATH/queries/scripts/create_table.py" \
-            --asterixdb-server localhost:19002 \
-            --hdfs-server hdfs://namenode:8020 \
-            --external-path "/$dataset_name/*.json.gz" \
-            --dataset-name Run2012B_SingleMu_${n}_untyped_internal \
-            --datatype any --file-format json --storage-location internal \
-            --log-level INFO --asterixdb-dataverse IrisHepBenchmark
+            # Internal relations loaded from files on HDFS
+            "$SCRIPT_PATH/queries/scripts/create_table.py" \
+                --asterixdb-server localhost:19002 \
+                --hdfs-server hdfs://namenode:8020 \
+                --external-path "/$dataset_name/*.json.gz" \
+                --dataset-name Run2012B_SingleMu_${n}_untyped_internal \
+                --datatype any --file-format json --storage-location internal \
+                --log-level INFO --asterixdb-dataverse IrisHepBenchmark
 
-        "$SCRIPT_PATH/queries/scripts/create_table.py" \
-            --asterixdb-server localhost:19002 \
-            --hdfs-server hdfs://namenode:8020 \
-            --external-path "/$dataset_name/*.json.gz" \
-            --dataset-name Run2012B_SingleMu_${n}_typed_internal \
-            --datatype event --file-format json --storage-location internal \
-            --log-level INFO --asterixdb-dataverse IrisHepBenchmark
+            "$SCRIPT_PATH/queries/scripts/create_table.py" \
+                --asterixdb-server localhost:19002 \
+                --hdfs-server hdfs://namenode:8020 \
+                --external-path "/$dataset_name/*.json.gz" \
+                --dataset-name Run2012B_SingleMu_${n}_typed_internal \
+                --datatype event --file-format json --storage-location internal \
+                --log-level INFO --asterixdb-dataverse IrisHepBenchmark
 
-        # External relations using files on HDFS
-        "$SCRIPT_PATH/queries/scripts/create_table.py" \
-            --asterixdb-server localhost:19002 \
-            --hdfs-server hdfs://namenode:8020 \
-            --external-path "/$dataset_name/*.json.gz" \
-            --dataset-name Run2012B_SingleMu_${n}_untyped_json_hdfs \
-            --datatype any --file-format json --storage-location external \
-            --log-level INFO --asterixdb-dataverse IrisHepBenchmark
+            # External relations using files on HDFS
+            "$SCRIPT_PATH/queries/scripts/create_table.py" \
+                --asterixdb-server localhost:19002 \
+                --hdfs-server hdfs://namenode:8020 \
+                --external-path "/$dataset_name/*.json.gz" \
+                --dataset-name Run2012B_SingleMu_${n}_untyped_json_hdfs \
+                --datatype any --file-format json --storage-location external \
+                --log-level INFO --asterixdb-dataverse IrisHepBenchmark
 
-        "$SCRIPT_PATH/queries/scripts/create_table.py" \
-            --asterixdb-server localhost:19002 \
-            --hdfs-server hdfs://namenode:8020 \
-            --external-path "/$dataset_name/*.json.gz" \
-            --dataset-name Run2012B_SingleMu_${n}_typed_json_hdfs \
-            --datatype event --file-format json --storage-location external \
-            --log-level INFO --asterixdb-dataverse IrisHepBenchmark
+            "$SCRIPT_PATH/queries/scripts/create_table.py" \
+                --asterixdb-server localhost:19002 \
+                --hdfs-server hdfs://namenode:8020 \
+                --external-path "/$dataset_name/*.json.gz" \
+                --dataset-name Run2012B_SingleMu_${n}_typed_json_hdfs \
+                --datatype event --file-format json --storage-location external \
+                --log-level INFO --asterixdb-dataverse IrisHepBenchmark
 
-        # External relations using files on S3
-        "$SCRIPT_PATH/queries/scripts/create_table.py" \
-            --asterixdb-server localhost:19002 \
-            --bucket-name $S3_INPUT_BUCKET \
-            --bucket-region $S3_REGION \
-            --secret-access-key $S3_SECRET_KEY \
-            --access-key-id $S3_ACCESS_KEY \
-            --external-path "$dataset_name/" \
-            --dataset-name Run2012B_SingleMu_${n}_untyped_json_s3 \
-            --datatype any --file-format json --storage-location external \
-            --log-level INFO --asterixdb-dataverse IrisHepBenchmark
+            # External relations using files on S3
+            "$SCRIPT_PATH/queries/scripts/create_table.py" \
+                --asterixdb-server localhost:19002 \
+                --bucket-name $S3_INPUT_BUCKET \
+                --bucket-region $S3_REGION \
+                --secret-access-key $S3_SECRET_KEY \
+                --access-key-id $S3_ACCESS_KEY \
+                --external-path "$dataset_name/" \
+                --dataset-name Run2012B_SingleMu_${n}_untyped_json_s3 \
+                --datatype any --file-format json --storage-location external \
+                --log-level INFO --asterixdb-dataverse IrisHepBenchmark
 
-        "$SCRIPT_PATH/queries/scripts/create_table.py" \
-            --asterixdb-server localhost:19002 \
-            --bucket-name $S3_INPUT_BUCKET \
-            --bucket-region $S3_REGION \
-            --secret-access-key $S3_SECRET_KEY \
-            --access-key-id $S3_ACCESS_KEY \
-            --external-path "$dataset_name/" \
-            --dataset-name Run2012B_SingleMu_${n}_typed_json_s3 \
-            --datatype event --file-format json --storage-location external \
-            --log-level INFO --asterixdb-dataverse IrisHepBenchmark
+            "$SCRIPT_PATH/queries/scripts/create_table.py" \
+                --asterixdb-server localhost:19002 \
+                --bucket-name $S3_INPUT_BUCKET \
+                --bucket-region $S3_REGION \
+                --secret-access-key $S3_SECRET_KEY \
+                --access-key-id $S3_ACCESS_KEY \
+                --external-path "$dataset_name/" \
+                --dataset-name Run2012B_SingleMu_${n}_typed_json_s3 \
+                --datatype event --file-format json --storage-location external \
+                --log-level INFO --asterixdb-dataverse IrisHepBenchmark
 
-        # External relations using local files
-        "$SCRIPT_PATH/queries/scripts/create_table.py" \
-            --asterixdb-server localhost:19002 \
-            --external-path "file:///data/input/$dataset_name/*.json.gz" \
-            --dataset-name Run2012B_SingleMu_${n}_untyped_json_local \
-            --datatype any --file-format json --storage-location external \
-            --log-level INFO --asterixdb-dataverse IrisHepBenchmark
+            # External relations using local files
+            "$SCRIPT_PATH/queries/scripts/create_table.py" \
+                --asterixdb-server localhost:19002 \
+                --external-path "file:///data/input/$dataset_name/*.json.gz" \
+                --dataset-name Run2012B_SingleMu_${n}_untyped_json_local \
+                --datatype any --file-format json --storage-location external \
+                --log-level INFO --asterixdb-dataverse IrisHepBenchmark
 
-        "$SCRIPT_PATH/queries/scripts/create_table.py" \
-            --asterixdb-server localhost:19002 \
-            --external-path "file:///data/input/$dataset_name/*.json.gz" \
-            --dataset-name Run2012B_SingleMu_${n}_typed_json_local \
-            --datatype event --file-format json --storage-location external \
-            --log-level INFO --asterixdb-dataverse IrisHepBenchmark
+            "$SCRIPT_PATH/queries/scripts/create_table.py" \
+                --asterixdb-server localhost:19002 \
+                --external-path "file:///data/input/$dataset_name/*.json.gz" \
+                --dataset-name Run2012B_SingleMu_${n}_typed_json_local \
+                --datatype event --file-format json --storage-location external \
+                --log-level INFO --asterixdb-dataverse IrisHepBenchmark
+            fi
 
         #
         # Parquet
